# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handlers.LoginEnvelope:
    properties:
      data:
        $ref: '#/definitions/services.LoginResult'
      success:
        example: true
        type: boolean
    type: object
  handlers.LoginRequest:
    properties:
      email:
        example: admin@mediguide.local
        type: string
      password:
        example: Admin123!
        type: string
    type: object
  handlers.LogoutEnvelope:
    properties:
      data:
        $ref: '#/definitions/handlers.LogoutResult'
      success:
        example: true
        type: boolean
    type: object
  handlers.LogoutResult:
    properties:
      logged_out:
        example: true
        type: boolean
    type: object
...
  handlers.PasswordChangeRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
      new_password_confirm:
        type: string
    type: object
  handlers.PasswordResetConfirmRequest:
    properties:
      password:
        type: string
      password_confirm:
        type: string
      token:
        type: string
    type: object
  handlers.PasswordResetRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
...
  handlers.RefreshRequest:
    properties:
      refresh_token:
        example: Gm8m3Wq2oJ7l6p4XnYx9QbT2f1WvL0H1v2z3k4m5n6o
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      address:
        example: Plot 12 Kampala Road
        type: string
      alternative_phone:
        example: "+256700000002"
        type: string
      avatar:
        example: https://example.com/avatar.png
        type: string
      city:
        example: Kampala
        type: string
      country:
        example: Uganda
        type: string
      department:
        example: Emergency
        type: string
      email:
        example: admin@mediguide.local
        type: string
      facility_id:
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        type: string
      job_title:
        example: Medical Officer
        type: string
      license_number:
        example: MD-12345
        type: string
      name:
        example: Admin User
        type: string
      notes:
        example: Night shift clinician
        type: string
      organization:
        example: Mulago Hospital
        type: string
      password:
        example: Admin123!
        type: string
      phone:
        example: "+256700000001"
        type: string
      postal_code:
        example: "256"
        type: string
      preferred_language:
        example: English
        type: string
      specialization:
        example:
        - Internal Medicine
        - Pediatrics
        items:
          type: string
        type: array
      timezone:
        example: Africa/Kampala
        type: string
    type: object
...
  handlers.UserEnvelope:
    properties:
      data:
        $ref: '#/definitions/models.User'
      success:
        example: true
        type: boolean
    type: object
...
  models.Permission:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      permissions_json:
        type: object
      role_key:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      address:
        type: string
      alternative_phone:
        type: string
      avatar:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      department:
        type: string
      email:
        type: string
      facility_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      job_title:
        type: string
      license_number:
        type: string
      name:
        type: string
      notes:
        type: string
      organization:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      preferred_language:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      specialization:
        items:
          type: string
        type: array
      status:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
      verified:
        type: boolean
    type: object
  services.AccountActionResult:
    properties:
      accepted:
        type: boolean
      delivery_accepted:
        type: boolean
      development_token:
        type: string
    type: object
...
  services.LoginResult:
    properties:
      expires_at:
        type: string
      refresh_expires_at:
        type: string
      refresh_token:
        type: string
      session_id:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
info:
  contact:
    email: admin@mediguide.local
    name: MediGuide Backend
  description: Offline-first clinical guideline backend API for MediGuide.
  license:
    name: Proprietary
  termsOfService: https://mediguide.local/terms
  title: MediGuide Backend API
  version: "1.0"
paths:
...
  /api/v2/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT bearer token.
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Log in a user
      tags:
      - auth
  /api/v2/auth/logout:
    post:
      description: Revoke the current authenticated session.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out current session
      tags:
      - auth
  /api/v2/auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset confirmation
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Confirm a password reset
      tags:
      - auth
  /api/v2/auth/password-reset/request:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.AccountActionResult'
      summary: Request a password reset
      tags:
      - auth
  /api/v2/auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a refresh token for a new access token and rotated refresh
        token.
      parameters:
      - description: Refresh payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Refresh a user session
      tags:
      - auth
  /api/v2/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new backend user account.
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.UserEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a user
      tags:
      - auth
...
  /api/v2/me:
    get:
      description: Return the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UserEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v2/me/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Password change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change the current user's password
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer access token. Example: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"